requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic

# iniciar con uvicorn app.main:app --reload
